Add temporary block for category mutation and refactored env vars

+++ src/modules/category/adapter/controller/category.controller.ts
@@ -1,8 +1,9 @@
-import { Body, Controller, Delete, Get, Param, Post, UseGuards } from '@nestjs/common';
+import { Body, Controller, Delete, Get, Param, Post, UnprocessableEntityException, UseGuards } from '@nestjs/common';
 import { CategoryService } from '../../usecase/service/category.service';
 import { SupabaseGuard } from 'src/modules/auth/guard/auth.guard';
 import { CategoryInput } from '../dto/category.input';
 import { CategoryOutput } from '../dto/category.output';
+import { ENABLE_CATEGORY_MUTATION } from 'src/modules/common/environment';
 
 @Controller('/categories')
 export class CategoryController {
@@ -11,12 +12,14 @@ export class CategoryController {
     @Post()
     @UseGuards(SupabaseGuard)
     async createCategory(@Body() category: CategoryInput): Promise<void> {
+        if (!ENABLE_CATEGORY_MUTATION) throw new UnprocessableEntityException('Category mutation is currently disabled');
         await this.categoryService.createCategory(category);
     }
 
     @Post('/:id')
     @UseGuards(SupabaseGuard)
     async updateCategory(@Param('id') id: string, @Body() category: CategoryInput): Promise<void> {
+        if (!ENABLE_CATEGORY_MUTATION) throw new UnprocessableEntityException('Category mutation is currently disabled');
         await this.categoryService.updateCategory(id, category);
     }
 
@@ -35,6 +38,7 @@ export class CategoryController {
     @Delete('/:id')
     @UseGuards(SupabaseGuard)
     async deleteCategory(@Param('id') id: string): Promise<void> {
+        if (!ENABLE_CATEGORY_MUTATION) throw new UnprocessableEntityException('Category mutation is currently disabled');
         await this.categoryService.deleteCategoryBy(id);
     }
 }
